+++ CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
   - Drop support for Ruby versions < 2.6
   - Added frozen string literals
   - Fix `SimpleTypedList#to_a` and `SimpleTypedList#to_ary` returning the internal list instance
+  - Remove ability to set `u=` to true in favor of using :single or one of the other underline options
 
 - **April.23.23**: 3.4.1
   - [PR #209](https://github.com/caxlsx/caxlsx/pull/209) - Revert characters other than `=` being considered as formulas.